we will need this different fixed/float encoding
and aMax is now a float too …

+++ ArduinoClient/types.h
@@ -20,7 +20,7 @@ struct movement {
   unsigned char motor;
   long target;
   double vMax;
-  long aMax;
+  double aMax;
   long jerk;
 };
